+++ src/app/pages/user/bucket/bucket-form-page/bucket-form-page.component.ts
@@ -6,7 +6,7 @@ import * as _ from 'lodash';
 import { Observable, of, timer } from 'rxjs';
 import { map, switchMap } from 'rxjs/operators';
 
-import { format } from '~/app/functions.helper';
+import { format, getErrorCode } from '~/app/functions.helper';
 import { DeclarativeFormComponent } from '~/app/shared/components/declarative-form/declarative-form.component';
 import { PageStatus } from '~/app/shared/components/page-status/page-status.component';
 import { DeclarativeFormConfig } from '~/app/shared/models/declarative-form-config.type';
@@ -22,6 +22,7 @@ export class BucketFormPageComponent implements OnInit {
   form!: DeclarativeFormComponent;
 
   public pageStatus: PageStatus = PageStatus.none;
+  public loadingErrorText: string = TEXT('Failed to load bucket.');
   public savingErrorText: string = TEXT('Failed to save bucket.');
   public config: DeclarativeFormConfig = {
     fields: []
@@ -48,8 +49,11 @@ export class BucketFormPageComponent implements OnInit {
           this.pageStatus = PageStatus.ready;
           this.form.patchValues(bucket);
         },
-        error: () => {
+        error: (err) => {
           this.pageStatus = PageStatus.loadingError;
+          this.loadingErrorText = format(TEXT('Failed to load bucket (code={{ code }}).'), {
+            code: getErrorCode(err)
+          });
         }
       });
     });
@@ -114,8 +118,8 @@ export class BucketFormPageComponent implements OnInit {
       },
       error: (err) => {
         this.pageStatus = PageStatus.savingError;
-        this.savingErrorText = format(TEXT('Failed to save bucket (code={{ err.code }}).'), {
-          err
+        this.savingErrorText = format(TEXT('Failed to save bucket (code={{ code }}).'), {
+          code: getErrorCode(err)
         });
       }
     });
@@ -129,8 +133,8 @@ export class BucketFormPageComponent implements OnInit {
       },
       error: (err) => {
         this.pageStatus = PageStatus.savingError;
-        this.savingErrorText = format(TEXT('Failed to save bucket (code={{ err.code }}).'), {
-          err
+        this.savingErrorText = format(TEXT('Failed to save bucket (code={{ code }}).'), {
+          code: getErrorCode(err)
         });
       }
     });